+++ breeze-core/src/test/kotlin/com/windea/breezeframework/core/extensions/CollectionExtensionsKtTest.kt
@@ -118,9 +118,7 @@ class CollectionExtensionsKtTest {
 			listOf(111, 222, 333, listOf(444)),
 			mapOf("a" to listOf("a"), "b" to listOf("b"))
 		)
-		list.deepFlatten(-2).also { println(it) }
 		list.deepFlatten().also { println(it) }
-		list.deepFlatten(0).also { println(it) }
 		list.deepFlatten(1).also { println(it) }
 		list.deepFlatten(2).also { println(it) }
 		list.deepFlatten(3).also { println(it) }
@@ -141,5 +139,15 @@ class CollectionExtensionsKtTest {
 		assertTrue(listOf("a").isIterableOf<String>())
 		assertTrue(mapOf("a" to "a").isMapOf<String, String>())
 		assertTrue(sequenceOf("a").isSequenceOf<String>())
+	}
+
+	@Test
+	fun expandTest2() {
+		val list = listOf<Any?>(1, listOf(2, 3, 4), listOf(5, listOf(6)), 7)
+		val flatList = list.flatMap { if(it is List<*>) it else listOf(it) }
+		val extendList = expand<Any?>(list) { if(it is List<*>) it else listOf() }
+		println(list)
+		println(flatList)
+		println(extendList)
 	}
 }