+++ breeze-core/src/test/kotlin/com/windea/breezeframework/core/extensions/CoreExtensionsKtTest.kt
@@ -24,8 +24,8 @@ class CoreExtensionsKtTest {
 
 	@Test
 	fun nearestStackInfo() {
-		DELAY("delay!")
-		FIXME("fix me!")
+		DELAY()
+		FIXME()
 	}
 
 	@Test